Refactoring: move pen tracker out of workers
