+++ src/rmview/workers.py
@@ -29,6 +29,7 @@ BYTES_PER_PIXEL = 2
 
 log = logging.getLogger('rmview')
 
+
 class FBWSignals(QObject):
   onFatalError = pyqtSignal(Exception)
   onNewFrame = pyqtSignal(QImage)
@@ -67,7 +68,6 @@ class RFB(RFBClient):
 
   def updateRectangle(self, x, y, width, height, data):
     self.painter.drawImage(x,y,QImage(data, width, height, width * BYTES_PER_PIXEL, IMG_FORMAT))
-
 
 
 class RFBFactory(RFBFactory):
@@ -310,95 +310,3 @@ class FrameBufferWorker(QRunnable):
     self.factory.instance.keyEvent(key)
     # time.sleep(.1)
     self.factory.instance.keyEvent(key, 0)
-
-
-class PWSignals(QObject):
-  onFatalError = pyqtSignal(Exception)
-  onPenMove = pyqtSignal(int, int)
-  onPenPress = pyqtSignal()
-  onPenLift = pyqtSignal()
-  onPenNear = pyqtSignal()
-  onPenFar = pyqtSignal()
-
-LIFTED = 0
-PRESSED = 1
-
-
-class PointerWorker(QRunnable):
-
-  _stop = False
-
-  def __init__(self, ssh, path="/dev/input/event0", threshold=1000):
-    super(PointerWorker, self).__init__()
-    self.event = path
-    self.ssh = ssh
-    self.threshold = threshold
-    self.signals = PWSignals()
-
-  @pyqtSlot()
-  def pause(self):
-    self.signals.blockSignals(True)
-
-  @pyqtSlot()
-  def resume(self):
-    self.signals.blockSignals(False)
-
-  def stop(self):
-    self._penkill.write('\n')
-    self._stop = True
-
-  @pyqtSlot()
-  def run(self):
-    penkill, penstream, _ = self.ssh.exec_command('cat %s & { read ; kill %%1; }' % self.event)
-    self._penkill = penkill
-    new_x = new_y = False
-    state = LIFTED
-
-    while not self._stop:
-      try:
-        _, _, e_type, e_code, e_value = struct.unpack('2IHHi', penstream.read(16))
-      except struct.error:
-        return
-      except Exception as e:
-        log.error('Error in pointer worker: %s %s', type(e), e)
-        return
-
-      # decoding adapted from remarkable_mouse
-      if e_type == e_type_abs:
-
-
-        # handle x direction
-        if e_code == e_code_stylus_xpos:
-          x = e_value
-          new_x = True
-
-        # handle y direction
-        if e_code == e_code_stylus_ypos:
-          y = e_value
-          new_y = True
-
-        # handle draw
-        if e_code == e_code_stylus_pressure:
-          if e_value > self.threshold:
-            if state == LIFTED:
-              log.debug('PRESS')
-              state = PRESSED
-              self.signals.onPenPress.emit()
-          else:
-            if state == PRESSED:
-              log.debug('RELEASE')
-              state = LIFTED
-              self.signals.onPenLift.emit()
-
-        if new_x and new_y:
-          self.signals.onPenMove.emit(x, y)
-          new_x = new_y = False
-
-      if e_type == e_type_key and e_code == e_code_stylus_proximity:
-        if e_value == 0:
-          self.signals.onPenFar.emit()
-        else:
-          self.signals.onPenNear.emit()
-
-
-
